refact:  divide dashboard views into separate apps for better responsibility management
