+++ dashboard/templates/index.html
@@ -237,7 +237,7 @@
                                         <td>
                                             <button class="text-blue-500 flex items-center justify-center w-full h-full">
                                                 {% for appointment in appointments %}
-                                                    {% if appointment.appointment_date == next_consult_date.date and appointment.appointment_time|format_time == time %}
+                                                    {% if appointment.date == next_consult_date.date and appointment.time|format_time == time %}
                                                         <div class="filled-cell">
                                                             <div class="current-patient-info">
                                                                 <span class="current-patient-name">{{ appointment.patient.name }}</span>
@@ -313,7 +313,7 @@
                     
                     </div>
                     <div class="panel mb-5 xl:w-full" id="searchPanel">
-                        <form method="get" action="{% url 'search_appointments' %}">
+                        <form id="searchForm">
                             <div class="r">
                                 <h3 class="mb-5 font-bold text-lg text-center">Procurar agendamento</h3>
                                 <label for="search-date">Por data</label>
@@ -322,7 +322,7 @@
                                 <input id="search-name" type="text" name="search-name" class="form-input mb-2" placeholder="Nome do Paciente">
                             </div>
                             <div class="mt-4 flex justify-end">
-                                <button type="submit" class="btn btn-primary w-full gap-2 btn-blue" >
+                                <button type="submit" class="btn btn-primary w-full gap-2 btn-blue">
                                     Pesquisar
                                 </button>
                             </div>
@@ -338,33 +338,11 @@
                                         <th>Ações</th>
                                     </tr>
                                 </thead>
-                                <tbody>
-                                    {% if not se_appointments %}
-                                        <tr>
-                                            <td colspan="4" class="text-center">Nenhum agendamento encontrado</td>
-                                        </tr>
-                                    {% else %}
-                                        {% for se_appointment in se_appointments %}
-                                        <tr>
-                                            <td class="text-center appointment-date">{{ se_appointment.appointment_date|date:"d/m/y" }}</td>
-                                            <td class="text-center appointment-time">{{ se_appointment.appointment_time|time:"H:i"  }}</td>
-                                            <td class="patient-name">{{ se_appointment.patient.name }}</td>
-                                            <td class="text-center">
-                                                <button onclick="showCancelAlert({{ se_appointment.id }})" class="cancel-appointment">
-                                                    <svg class="btn-svg" fill="#000000" width="800px" height="800px" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
-                                                        <path d="M5.755,20.283,4,8H20L18.245,20.283A2,2,0,0,1,16.265,22H7.735A2,2,0,0,1,5.755,20.283ZM21,4H16V3a1,1,0,0,0-1-1H9A1,1,0,0,0,8,3V4H3A1,1,0,0,0,3,6H21a1,1,0,0,0,0-2Z"/>
-                                                    </svg>
-                                                </button>
-                                            </td>
-                                        </tr>
-                                        {% endfor %}
-                                    {% endif %}
+                                <tbody id="appointmentsTableBody">
                                 </tbody>
                             </table>
                         </div>
-                    </div>
-                    
-                    
+                    </div>                    
                 </div>
             </div>
         </div>
