+++ patients/admin.py
@@ -1,2 +1,3 @@
 from django.contrib import admin
 
+# Register your models here.
